# Improve performance upper/lower by executing in place and cpu dynamic dispatch