--- a/src/Functions/FunctionStringToString.h
+++ b/src/Functions/FunctionStringToString.h
@@ -1,11 +1,13 @@
 #pragma once
 
-#include <DataTypes/DataTypeString.h>
-#include <Columns/ColumnString.h>
 #include <Columns/ColumnFixedString.h>
+#include <Columns/ColumnString.h>
+#include <DataTypes/DataTypeString.h>
 #include <Functions/FunctionHelpers.h>
 #include <Functions/IFunction.h>
+#include <Functions/LowerUpperImpl.h>
 #include <Interpreters/Context_fwd.h>
+#include <iostream>
 
 
 namespace DB
@@ -61,6 +63,30 @@ public:
 
     ColumnPtr executeImpl(const ColumnsWithTypeAndName & arguments, const DataTypePtr &, size_t /*input_rows_count*/) const override
     {
+        if constexpr (IsLowerUpper<Impl>::value)
+        {
+            // std::cout << "is lower upper:" << arguments[0].column->use_count() << std::endl;
+            // std::cout << StackTrace().toString() << std::endl;
+            if (arguments[0].column->use_count() == 1)
+            {
+                // std::cout << "execute in place" << std::endl;
+                MutableColumnPtr mutable_column = arguments[0].column->assumeMutable();
+                if (ColumnString * col = typeid_cast<ColumnString *>(mutable_column.get()))
+                {
+                    Impl::vectorInPlace(col->getChars(), col->getOffsets());
+                    return std::move(mutable_column);
+                }
+                else if (ColumnFixedString * col_fixed = typeid_cast<ColumnFixedString *>(mutable_column.get()))
+                {
+                    Impl::vectorFixedInPlace(col_fixed->getChars(), col_fixed->getN());
+                    return std::move(mutable_column);
+                }
+                else
+                    throw Exception(
+                        ErrorCodes::ILLEGAL_COLUMN, "Illegal column {} of argument of function {}", mutable_column->getName(), getName());
+            }
+        }
+
         const ColumnPtr column = arguments[0].column;
         if (const ColumnString * col = checkAndGetColumn<ColumnString>(column.get()))
         {