### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Improve performance upper/lower by executing in place and cpu dynamic dispatch 

TODO: apply it for other string functions like `initcap` and `substring`


A simple test shows 1.49x speedup. 
```
Q:  select lower(materialize('hello world')) from numbers(100000000) format Null;

baseline: 
0 rows in set. Elapsed: 0.484 sec. Processed 100.00 million rows, 800.00 MB (206.48 million rows/s., 1.65 GB/s.)
Peak memory usage: 4.51 MiB. 

opt1: execute lower/upper in place if needed
0 rows in set. Elapsed: 0.397 sec. Processed 100.00 million rows, 800.00 MB (251.89 million rows/s., 2.02 GB/s.)
Peak memory usage: 4.50 MiB. 

opt2: opt1 + cpu dynamic dispatch 
0 rows in set. Elapsed: 0.324 sec. Processed 100.00 million rows, 800.00 MB (308.31 million rows/s., 2.47 GB/s.)
Peak memory usage: 4.50 MiB. 
```