@rschu1ze 2024-05-06
I split out the vectorization bits into #63428, omitting the controversial in-place updates, but then had this realization:
- in the previous code, starting with the input string prefix, we processed as many 512 bit chunks as possible with AVX-512, then as many as possible 256 bit chunks with AVX2 (<-- actually not implemented but could be easily implemented), then as many as possible 128 bit chunks with SSE, and finally the tail was handled sequentially. Interestingly, ClickHouse is compiled by default with SSE instructions, meaning that we would only do SSE + generic processing.
- in this PR, dynamic dispatch is used based on the capabilities of the target system (AVX-512, AVX2, SSE, generic).

The problem is that in practice most strings are short, e.g. shorter than 128 bits. As a result, the new code would call the AVX-512 kernel (for example) but then do tail processing (which would be compiler-auto-vectorized but it is still slow). To solve that, the individual kernels would need to apply a similar strategy as the previous code, i.e. the AVX-512 kernel would process as much data as possible with AVX-512, then with AVX-2, then SSE and finally handle the tail.

But I am not sure if all of this trouble is really worth the effort, as we are really optimizing for an unlikely use case (very long strings) while trying to not break performance of the common case (short strings).

As a result, I closed my other PR and I guess it makes sense to discontinue this PR as well.